Feat: 추가 Club 조회 뷰 내부 정보 카운팅 기능 ClubUserCountInfoDto 참조

+++ src/main/java/mangpo/server/repository/CommentRepository.java
@@ -12,5 +12,5 @@ public interface CommentRepository extends JpaRepository<Comment,Long> {
     List<Comment> findAllByPost(Post post);
     void deleteAllByUser(User user);
 
-    Long countByPost(Post post);
+    Integer countByPost(Post post);
 }
