+++ src/main/java/mangpo/server/service/LikedService.java
@@ -42,7 +42,7 @@ public class LikedService {
         likedRepository.deleteAll(likedList);
     }
 
-    public Long countByPost(Post post){
+    public Integer countByPost(Post post){
         return likedRepository.countByPost(post);
     }
